+++ gui/src/fm.rs
@@ -143,7 +143,7 @@ pub fn fm() {
                     keycode: Some(keycode),
                     ..
                 } => {
-                    let mut set = |nn: u32| {
+                    let mut set = |nn: u8| {
                         osc_freq_ctrl
                             .lock()
                             .set_value_at_time(audio_time, notenum_to_frequency(nn));
